LPS-89891 Adds new management toolbar class for select layout page template collections
