+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/display/context/LayoutPageTemplateCollectionsManagementToolbarDisplayContext.java
@@ -1,0 +1,89 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.layout.admin.web.internal.display.context;
+
+import com.liferay.frontend.taglib.clay.servlet.taglib.display.context.SearchContainerManagementToolbarDisplayContext;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
+import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.WebKeys;
+
+import javax.portlet.PortletURL;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * @author Eudaldo Alonso
+ */
+public class LayoutPageTemplateCollectionsManagementToolbarDisplayContext
+	extends SearchContainerManagementToolbarDisplayContext {
+
+	public LayoutPageTemplateCollectionsManagementToolbarDisplayContext(
+		LiferayPortletRequest liferayPortletRequest,
+		LiferayPortletResponse liferayPortletResponse,
+		HttpServletRequest request,
+		LayoutPageTemplateCollectionsDisplayContext
+			layoutPageTemplateCollectionsDisplayContext) {
+
+		super(
+			liferayPortletRequest, liferayPortletResponse, request,
+			layoutPageTemplateCollectionsDisplayContext.getSearchContainer());
+
+		_request = request;
+	}
+
+	@Override
+	public String getClearResultsURL() {
+		PortletURL clearResultsURL = getPortletURL();
+
+		clearResultsURL.setParameter("keywords", StringPool.BLANK);
+
+		return clearResultsURL.toString();
+	}
+
+	@Override
+	public String getSearchActionURL() {
+		ThemeDisplay themeDisplay = (ThemeDisplay)_request.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		PortletURL searchActionURL = liferayPortletResponse.createRenderURL();
+
+		searchActionURL.setParameter(
+			"mvcRenderCommandName",
+			"/layout/select_layout_page_template_collections");
+		searchActionURL.setParameter("redirect", themeDisplay.getURLCurrent());
+
+		return searchActionURL.toString();
+	}
+
+	@Override
+	public String getSearchContainerId() {
+		return "layoutPageTemplateCollections";
+	}
+
+	@Override
+	protected String[] getNavigationKeys() {
+		return new String[] {"all"};
+	}
+
+	@Override
+	protected String[] getOrderByKeys() {
+		return new String[] {"create-date", "name"};
+	}
+
+	private final HttpServletRequest _request;
+
+}
